LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-createOffer-expected.txt


PASS createOffer() returns RTCSessionDescriptionInit
PASS createOffer() and then setLocalDescription() should succeed
PASS createOffer() after connection is closed should reject with InvalidStateError
PASS When media stream is added when createOffer() is running in parallel, the result offer should contain the new media stream
PASS createOffer() should fail when signaling state is not stable or have-local-offer
